Proxies - Add App Proxy (Yaml) - 9424791148 (#110)
* Proxies - App - Initial Commit

* Proxies - App - Add Yaml Tests

* Repos - App - Update Content

* Repos - Remove App Tests
